requirements.txt
biopython==1.70